refactor: Dao 인터페이스 도입
DAO 의 기능명세를 위한 인터페이스를 도입하였습니다.
