+++ src/test/java/wooteco/subway/dao/JdbcStationDaoTest.java
@@ -9,16 +9,17 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
+import wooteco.subway.dao.jdbc.JdbcStationDao;
 import wooteco.subway.domain.Station;
 
 @JdbcTest
-class StationDaoTest {
+class JdbcStationDaoTest {
 
-    private final StationDao stationDao;
+    private final JdbcStationDao jdbcStationDao;
 
     @Autowired
-    public StationDaoTest(JdbcTemplate jdbcTemplate) {
-        this.stationDao = new StationDao(jdbcTemplate);
+    public JdbcStationDaoTest(JdbcTemplate jdbcTemplate) {
+        this.jdbcStationDao = new JdbcStationDao(jdbcTemplate);
     }
 
     @Test
@@ -28,10 +29,10 @@ class StationDaoTest {
         final Station station = new Station("지하철역이름");
 
         // when
-        final Long savedId = stationDao.save(station);
+        final Long savedId = jdbcStationDao.save(station);
 
         // then
-        final Station savedStation = stationDao.findById(savedId);
+        final Station savedStation = jdbcStationDao.findById(savedId);
         assertThat(station).isEqualTo(savedStation);
     }
 
@@ -43,12 +44,12 @@ class StationDaoTest {
         final Station station2 = new Station("새로운지하철역이름");
         final Station station3 = new Station("또다른지하철역이름");
 
-        stationDao.save(station1);
-        stationDao.save(station2);
-        stationDao.save(station3);
+        jdbcStationDao.save(station1);
+        jdbcStationDao.save(station2);
+        jdbcStationDao.save(station3);
 
         // when
-        final List<Station> stations = stationDao.findAll();
+        final List<Station> stations = jdbcStationDao.findAll();
 
         // then
         assertThat(stations).hasSize(3)
@@ -61,9 +62,9 @@ class StationDaoTest {
     void deleteById() {
         // given
         final Station station = new Station("지하철역이름");
-        final Long savedId = stationDao.save(station);
+        final Long savedId = jdbcStationDao.save(station);
 
         // when & then
-        assertDoesNotThrow(() -> stationDao.deleteById(savedId));
+        assertDoesNotThrow(() -> jdbcStationDao.deleteById(savedId));
     }
 }
